add fallback procedure for swap api
